Cleaning up documentation, test files

+++ demo.sh
@@ -1,7 +1,9 @@
 #!/bin/bash
 
-alias stap=~/systemtap-4.0/stap
+alias stap=../systemtap-4.0/stap
 cd systemtap_scripts
+
+file_path=/home/andre/OS-Runtime-Verification/systemtap_scripts/kmalloc_app
 
 if [ "$1" = "clean" ]
 then
@@ -24,7 +26,7 @@ else
 	then
 		# Set up instrumentation, give it some time
 		sudo ./dummy &
-		sudo stap -g offline/kmalloc_instrumentation.stp /home/andre/OS-Runtime-Verification/systemtap_scripts/kmalloc_app kmalloc_app &
+		sudo stap -g offline/kmalloc_instrumentation.stp $file_path kmalloc_app &
 		sleep 6
 		echo "INSTRUMENTATION SETUP COMPLETED"
 		#Run test program
@@ -43,23 +45,26 @@ else
 		sleep 2
 		echo "OFFLINE MONITORING COMPLETED"
 	else
-		# Set up instrumentation with online monitor, give it some time
-		cd online
-		sudo dmesg -c &> /dev/null
-		sudo insmod monitor_lib.ko
-		sudo insmod monitor_spec1.ko
-		sudo insmod monitor_spec2.ko
-		sudo insmod monitor_spec3.ko
-		sudo ../dummy &
-		sudo stap -g kmalloc_instrumentation.stp /home/andre/OS-Runtime-Verification/systemtap_scripts/kmalloc_app kmalloc_app &
-		sleep 10
-		echo "INSTRUMENTATION SETUP COMPLETED"
-		echo "ONLINE MONITORING ENGAGED"
-		#Run test program
-		echo "RUNNING TEST FILE"
-		sudo time ../kmalloc_app
-		sleep 3
-		sudo killall dummy
-		echo "ONLINE MONITORING COMPLETED"
+		if [[ "$1" == "online" ]]
+		then
+			# Set up instrumentation with online monitor, give it some time
+			cd online
+			sudo dmesg -c &> /dev/null
+			sudo insmod monitor_lib.ko
+			sudo insmod monitor_spec1.ko
+			sudo insmod monitor_spec2.ko
+			sudo insmod monitor_spec3.ko
+			sudo ../dummy &
+			sudo stap -g kmalloc_instrumentation.stp $file_path kmalloc_app &
+			sleep 10
+			echo "INSTRUMENTATION SETUP COMPLETED"
+			echo "ONLINE MONITORING ENGAGED"
+			#Run test program
+			echo "RUNNING TEST FILE"
+			sudo time ../kmalloc_app
+			sleep 3
+			sudo killall dummy
+			echo "ONLINE MONITORING COMPLETED"
+		fi
 	fi
 fi
